+++ src/mlinstruct/train/trainer/__init__.py
@@ -2,10 +2,11 @@ from typing import Any
 
 from mlinstruct.train.trainer.base_trainer import BaseTrainer
 from mlinstruct.train.trainer.default_trainer import DefaultTrainer
+from mlinstruct.train.trainer.epoch_loop_trainer import EpochLoopTrainer
 from mlinstruct.train.trainer.kfold_trainer import KFoldTrainer
 from mlinstruct.utils.optional_deps import require
 
-__all__ = ["BaseTrainer", "DefaultTrainer", "KFoldTrainer", "GANTrainer"]
+__all__ = ["BaseTrainer", "EpochLoopTrainer", "DefaultTrainer", "KFoldTrainer", "GANTrainer"]
 
 
 def __getattr__(name: str) -> Any: